Simpler build system and replacing `Vec::extract_if` with `Vec::retain`

+++ src/loader.rs
@@ -810,7 +810,7 @@ impl Default for Symbols {
 /// If everything went as expected, the program should crash in a minute or two:
 ///
 /// ```text
-/// lyte@mini ~/simple_fuzzer > CERT_KEYCHAIN=Impalabs make run
+/// lyte@mini ~/simple_fuzzer > make run
 /// # [...]
 /// Loading corpus...
 /// Corpus loaded!
@@ -1003,10 +1003,7 @@ pub trait Loader: Clone + Send {
         let scaled_max_mutations = max_mutations as f64 / max_size as f64 * data.len() as f64;
         let scaled_max_mutations = std::cmp::max(2, scaled_max_mutations as u64);
         // It's safe to unwrap since 0 < scaled_max_mutations.
-        let nb_mutations = mutator
-            .rand
-            .u64_range(1, scaled_max_mutations as u64)
-            .unwrap();
+        let nb_mutations = mutator.rand.u64_range(1, scaled_max_mutations).unwrap();
         // Randomly mutates the input.
         for _ in 0..nb_mutations {
             let strat_idx = mutator.rand.u64();
